Revert "add redact util functions and upgrade go version (#75)"
This reverts commit 6657a5023d8a3729d10d8659bce4df267186b585.
