+++ go.mod
@@ -1,14 +1,8 @@
 module github.com/pingcap/errors
 
-go 1.23
+go 1.14
 
 require (
 	github.com/stretchr/testify v1.4.0
 	go.uber.org/atomic v1.6.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v2 v2.2.2 // indirect
 )
